Refactor/node names (#3)
* Pimplified ScheduleDataNode


* Pimplified TrajectoryServer


* Fixed segfault in TrajectoryServer by using reference wrapper


* Fix segfault in parse_trajectories


* Rename target from rviz2 to schedule_visualizer. Update visualizer launch file


* Cleanup and uncrustify


* Reorder headers, move ScheduleMarkerPublisher into separate file, other feedback
