+++ rmf_schedule_visualizer/test/CMakeLists.txt
@@ -1,16 +1,16 @@
 
 find_package(Eigen3 REQUIRED)
 
-add_executable(submit_trajectory SubmitTrajectory.cpp)
+add_executable(test_submit_trajectory SubmitTrajectory.cpp)
 
-target_link_libraries(submit_trajectory
+target_link_libraries(test_submit_trajectory
   PUBLIC
     rmf_traffic::rmf_traffic
     rmf_traffic_ros2::rmf_traffic_ros2
     ${rmf_traffic_msgs_LIBRARIES}
 )
 
-target_include_directories(submit_trajectory
+target_include_directories(test_submit_trajectory
   PRIVATE
     $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/include>
   PUBLIC
@@ -18,7 +18,7 @@ target_include_directories(submit_trajectory
 )
 
 install(
-  TARGETS submit_trajectory
-  EXPORT submit_trajectory
+  TARGETS test_submit_trajectory
+  EXPORT test_submit_trajectory
   RUNTIME DESTINATION lib/${PROJECT_NAME}
 )
