+++ visualizer/launch/visualizer.launch.xml
@@ -1,0 +1,33 @@
+<?xml version = '1.0' ?>
+
+<launch>
+
+  <arg name="use_sim_time" default="false"/>
+  <arg name="rate" default="10"/>
+  <arg name="map_name" default="B1"/>
+  <arg name="viz_config_file" default="$(find-pkg-share rmf_schedule_visualizer)/config/rmf.rviz"/>
+  <arg name="display_names" default="true"/>
+  <arg name="websocket_port" default="8006"/>
+  <arg name="headless" default="false" description="Do not launch rviz2"/>
+
+  <group>
+    <node pkg="rmf_schedule_visualizer" exec="schedule_visualizer" args="-r $(var rate) -m $(var map_name) -p $(var websocket_port) --history 50">
+      <param name="use_sim_time" value="$(var use_sim_time)"/>
+    </node>
+
+    <node pkg="fleet_state_visualizer" exec="fleet_state_visualizer" args="-m $(var map_name)">
+      <param name="display_names" value ="$(var display_names)"/>
+      <param name="use_sim_time" value="$(var use_sim_time)"/>
+    </node>
+
+    <node pkg="building_systems_visualizer" exec="building_systems_visualizer" args="-m $(var map_name)">
+      <param name="use_sim_time" value="$(var use_sim_time)"/>
+      <param name="use_sim_time" value="$(var use_sim_time)"/>
+    </node>
+  </group>
+
+  <group unless="$(var headless)">
+    <executable cmd="rviz2 -d $(var viz_config_file)" output="both"/>
+  </group>
+
+</launch>
